feat: Introduce new event sections including BuildathonHero, BuildathonStats, BuildathonOverview, and BuildathonSponsors; refactor existing components for improved structure and clarity
