+++ components/event/speaker-section/speaker-headers.tsx
@@ -1,6 +1,7 @@
 import { ArrowLeft, ArrowRight } from "lucide-react";
 import StackGridButton from "../stack-grid-button";
 import { SpeakersHeaderProps } from "@/types/speaker.types";
+import StrokedText from "../stroked-text";
 
 export function SpeakersHeader({
   currentIndex,
@@ -10,15 +11,10 @@ export function SpeakersHeader({
 }: SpeakersHeaderProps) {
   return (
     <div className="px-4 lg:px-0 flex justify-between max-w-[1440px] mx-auto w-full mb-8">
-      <h2
+      <StrokedText
+        text="Speakers line up"
         className="lg:text-6xl text-4xl uppercase text-white font-extrabold"
-        style={{
-          textShadow:
-            "1px 1px 0px #000000, -1px -1px 0px #000000, 1px -1px 0px #000000, -1px 1px 0px #000000",
-        }}
-      >
-        Speakers line up
-      </h2>
+      />
 
       {/* Desktop Navigation */}
       <div className="lg:flex gap-4 items-center hidden">
